Solved nasty bugs after understanding lisp and raylib more
So, as it turns out, quoting anything creates a 'compile time list' in
lisp.

So, even using those inside of functions and later 'sharing' and
modifying would cause it to not be 'reset' for some reason.

And thus i was using 'copy-list'

But that was not enough, the nil values are also list, so they also
will become 'compile time' things and will be shared and not
're-evaluated' in the list formed by quoted.

So it turns out, you have to use 'copy-tree' to avoid such problems

[I still dont know entirely why this bug occured though]

And also, it turned out, I have to 'recreate' the font on each app
run. So I added it to the 'destructor'.